fix: Prevent passing a NULL pointer to strncmp when creating an empty string
While trying to find a conststr for a zero-length (empty) it might happen that
the string pointer is NULL (for example during parsing a JSON string).
This caused berry to pass the NULL pointer to strncmp which is Undefined
Behaviour, and caused the address sanitizer to print a warning (also this
could cause crashes on some configurations).

Additionally a test case has been added for this behaciour.