feature: Make parsing JSONnumbers more performant & compliant with the JSON spec
Added a new function for parsing JSON numbers: "parser_number". In
contrast with the built in `be_str2num` it is more restrictive (doesn't allow
hexadecimal numbers, leading zeros, dots without the fractional part.

It also improves the performance of parsing a little bit, because it doesn't
reparse the integer part when it encounters a real number.