tests: Add advanced tests for the json parser
Introduced a test which checks the json parser against some pathological
json strings, to ensure it doesn't crash and is conformant with the JSON spec.

The test data was adapted from https://github.com/nst/JSONTestSuite

The test includes both positive cases, which should be successfully parsed and
negative ones which should be rejected.

Also included are "any" cases, whcich the parser is free to accept or reject,
but it shouldn't crash the whole interpreter.